requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.19
python-dotenv==1.0.0
stripe==5.5.0
requests==2.31.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0